+++ space.js
@@ -130,27 +130,24 @@ Space.fromFile = function(filepath, options, callback) {
 
 /**
  * @param str string The csv string to parse
- * @param propertyName? string Optional property name to set for each row. Default is "row"
  * @return space
  */
-Space.fromCsv = function (str, propertyName) {
-  return Space.fromDelimiter(str, ",", propertyName)
+Space.fromCsv = function (str) {
+  return Space.fromDelimiter(str, ",")
 }
 
 /**
  * @param str string The csv string to parse
  * @param delimiter string
- * @param propertyName? string Optional property name to set for each row. Default is "row"
  * @return space
  */
-Space.fromDelimiter = function (str, delimiter, propertyName) {
+Space.fromDelimiter = function (str, delimiter) {
   var length = str.length,
       currentItem = "",
       currentPosition = 0,
       inQuote = false,
       rows = [[]],
       space = new Space(),
-      propertyName = propertyName || "row",
       currentRow = 0
 
   while (currentPosition < length) {
@@ -211,7 +208,8 @@ Space.fromDelimiter = function (str, delimiter, propertyName) {
       obj.append(prop, v)
     })
 
-    space.append(propertyName, obj)
+    // Subtract 1 since header was row 0
+    space.append(index - 1, obj)
   })
 
   return space
@@ -221,20 +219,18 @@ Space.fromDelimiter = function (str, delimiter, propertyName) {
  * Parses a simple space separated value "name age height\njoe 20 68"
  *
  * @param str string ssv string to parse
- * @param propertyName? string Optional property name to set for each row. Default is "row"
  * @return space
  */
-Space.fromSsv = function (str, propertyName) {
-  return Space.fromDelimiter(str, " ", propertyName)
+Space.fromSsv = function (str) {
+  return Space.fromDelimiter(str, " ")
 }
 
 /**
  * @param str string The tab string to parse
- * @param propertyName? string Optional property name to set for each row. Default is "row"
  * @return space
  */
-Space.fromTsv = function (str, propertyName) {
-  return Space.fromDelimiter(str, "\t", propertyName)
+Space.fromTsv = function (str) {
+  return Space.fromDelimiter(str, "\t")
 }
 
 /**
@@ -2285,6 +2281,36 @@ Space.prototype.toTsv = function() {
  */
 Space.prototype.toURL = function() {
   return encodeURIComponent(this.toString())
+}
+
+/**
+ * @param pretty? boolean 
+ * @return string
+ */
+Space.prototype.toXML = function(pretty) {
+  return this._toXML(pretty ? 0 : -1)
+}
+
+Space.prototype._toXML = function(spaceCount) {
+  var xml = "",
+      spaces = spaceCount === -1 ? "" : Space.strRepeat(" ", spaceCount)
+  
+  this.each(function(property, value) {
+    xml += spaces + "<" + property + ">"
+    if (value instanceof Space) {
+      if (value.length() === 0) {
+        xml += "</" + property + ">" + (spaceCount === -1 ? "" : "\n")
+      } else {
+          xml += (spaceCount === -1 ? "" : "\n") + value._toXML(spaceCount > -1 ? spaceCount + 2 : -1)
+          xml += spaces + "</" + property + ">" + (spaceCount === -1 ? "" : "\n")
+        }
+    }
+    else {
+      xml += value
+      xml += "</" + property + ">" + (spaceCount === -1 ? "" : "\n")
+    }
+  })
+  return xml
 }
 
 Space.prototype.__transpose = function(templateString) {